+++ handlers/grpcServiceHandler.go
@@ -85,17 +85,15 @@ func (gsh *GRPCServiceHandler) CreateTenant(ctx context.Context, tenantMeta *pb.
 	threshProfileResponse, err := gsh.tsh.CreateTenantThresholdProfile(ctx, &threshPrfReq)
 	if err != nil {
 		logger.Log.Errorf("Unable to create Threshold Profile for Tenant %s. The Tenant does exist though, so may need to create the Threshold Profile manually", result.GetXId())
-	} else {
-		// Update the tenant metadata with the new default threshold profile ID.
-		tenantUpdateRequest := pb.TenantDescriptorRequest{XId: result.GetXId(), XRev: result.GetXRev(), Data: result.GetData()}
-		tenantUpdateRequest.GetData().DefaultThresholdProfile = threshProfileResponse.GetXId()
-		resultWithThreshProfile, err := gsh.ash.UpdateTenantDescriptor(ctx, &tenantUpdateRequest)
-		if err != nil {
-			logger.Log.Errorf("Unable to assign Threshold Profile %s to Tenant %s. May need to assign the Threshold Profile manually", threshProfileResponse.GetXId(), result.GetXId())
-			return result, nil
-		}
+	}
 
-		return resultWithThreshProfile, nil
+	// Create the tenant metadata
+	meta := createDefaultTenantMeta(result.GetXId(), threshProfileResponse.GetXId(), result.GetData().GetName())
+	metaID := db.GenerateID(meta, string(db.TenantMetaType))
+	metaReq := pb.TenantMeta{XId: metaID, Data: meta}
+	_, err = gsh.tsh.CreateTenantMeta(ctx, &metaReq)
+	if err != nil {
+		logger.Log.Errorf("Unable to create Tenant Meta for Tenant %s. The Tenant does exist though, so may need to create the Tenant Meta manually", result.GetXId())
 	}
 
 	return result, nil
@@ -300,4 +298,24 @@ func (gsh *GRPCServiceHandler) GetThresholdCrossingByMonitoredObject(ctx context
 func (gsh *GRPCServiceHandler) GetHistogram(ctx context.Context, histogramReq *pb.HistogramRequest) (*pb.JSONAPIObject, error) {
 
 	return gsh.msh.GetHistogram(ctx, histogramReq)
+}
+
+// CreateTenantMeta - Create TenantMeta scoped to a Single Tenant.
+func (gsh *GRPCServiceHandler) CreateTenantMeta(ctx context.Context, meta *pb.TenantMeta) (*pb.TenantMeta, error) {
+	return gsh.tsh.CreateTenantMeta(ctx, meta)
+}
+
+// UpdateTenantMeta - Update TenantMeta scoped to a single Tenant.
+func (gsh *GRPCServiceHandler) UpdateTenantMeta(ctx context.Context, meta *pb.TenantMeta) (*pb.TenantMeta, error) {
+	return gsh.tsh.UpdateTenantMeta(ctx, meta)
+}
+
+// DeleteTenantMeta - Delete TenantMeta scoped to a single Tenant.
+func (gsh *GRPCServiceHandler) DeleteTenantMeta(ctx context.Context, tenantID *wr.StringValue) (*pb.TenantMeta, error) {
+	return gsh.tsh.DeleteTenantMeta(ctx, tenantID)
+}
+
+// GetTenantMeta - Retrieve a User scoped to a single Tenant.
+func (gsh *GRPCServiceHandler) GetTenantMeta(ctx context.Context, tenantID *wr.StringValue) (*pb.TenantMeta, error) {
+	return gsh.tsh.GetTenantMeta(ctx, tenantID)
 }