+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Current Release 
+#### adh-gather:0.27.0
+* feat - adding the Tenant Meta data model to tenant DB.
 ### 0.26.0
 * Fixes for SLA Domain Report 
 ### 0.25.0